Arrange files and components for CreateRole page.
