+++ public/apps/configuration-react/pages/CreateRole/components/ClusterPermissions/index.js
@@ -1,0 +1,2 @@
+import ClusterPermissions from './ClusterPermissions';
+export default ClusterPermissions;
