+++ public/apps/configuration-react/pages/CreateRole/components/IndexPermissions/index.js
@@ -1,0 +1,2 @@
+import IndexPermissions from './IndexPermissions';
+export default IndexPermissions;
