+++ public/apps/configuration-react/pages/CreateRole/components/Overview/index.js
@@ -1,0 +1,2 @@
+import Overview from './Overview';
+export default Overview;
