+++ public/apps/configuration-react/pages/CreateRole/components/TenantPermissions/index.js
@@ -1,0 +1,2 @@
+import TenantPermissions from './TenantPermissions';
+export default TenantPermissions;
